Yylib.symlink: use absolute paths when creating hardlinks.
The target of symlinks is resolved relative to the symlink path. With
hardlinks, it is resolved relative to the working directory at the time of
creation.
Using an absolute path built with ((dirname link) ^ "/" ^ target) is simple
and works for our case; it fails if target is an absolute path but there is
currently no package with such a link so it will be fine for now.