Yylib.symlink: don't prepend a '/' to the target of symlinks at the root.
The "i686-w64-mingw32" -> "mingw" symlink had its target changed to
"/mingw" on windows.
The transformation is done because symlinks targets are interpreted
relative to the location of the symlink while hardlinks and reparse points
are interpreted relative to the current working directory.
However the transformation was flawed and would preped a "/" for targets in
the root directory.