fileutils: patches to improve symlinks support as needed in win-builds.
This patches are against ocaml-fileutils.
The root of the issue is that fileutils has a type named "kind" which
can be "Directory", "Regular", ... but not "Symlink". In other words,
fileutils will always try to resolve symlinks and raise an exception for
broken symlinks. In yypkg, we need to have a "Symlink" variant for the
Windows support.

This has been sent to the fileutils bugtracker as preliminary work but
time has been too scarce to get them in a mergeable shape which has been
triple-checked. That said, I haven't found an issue with them yet; it's
simply that the patches break API and might break behaviour for others.

See the bugtracker:
https://forge.ocamlcore.org/tracker/index.php?func=detail&aid=712&group_id=128&atid=589